Move add_hyperlink to it's own file
