+++ resume_writer/resume_render/simple/simple_personal_section.py
@@ -1,90 +1,15 @@
 import logging
 
 import docx.document
-from docx import Document
 from docx.enum.text import WD_ALIGN_PARAGRAPH
 from docx.shared import Pt
-from docx.text.paragraph import Paragraph
+from resume_render.docx_hyperlink import add_hyperlink
 from resume_render.render_settings import ResumePersonalSettings
 from resume_render.resume_render_base import ResumeRenderPersonalBase
 
 from resume_writer.models.personal import ContactInfo, Personal
 
 log = logging.getLogger(__name__)
-
-
-def get_or_create_hyperlink_style(d : Document) -> str:
-    """Create a hyperlink style if one doesn't exist, or return existing style."""
-
-    """If this document had no hyperlinks so far, the builtin
-    Hyperlink style will likely be missing and we need to add it.
-    There's no predefined value, different Word versions
-    define it differently.
-    This version is how Word 2019 defines it in the
-    default theme, excluding a theme reference.
-    """
-    if "Hyperlink" not in d.styles:
-        if "Default Character Font" not in d.styles:
-            ds = d.styles.add_style(
-                "Default Character Font",
-                docx.enum.style.WD_STYLE_TYPE.CHARACTER,
-                True,  # noqa: FBT003
-            )
-            ds.element.set(docx.oxml.shared.qn("w:default"), "1")
-            ds.priority = 1
-            ds.hidden = True
-            ds.unhide_when_used = True
-            del ds
-        hs = d.styles.add_style(
-            "Hyperlink",
-            docx.enum.style.WD_STYLE_TYPE.CHARACTER,
-            True,  # noqa: FBT003
-        )
-        hs.base_style = d.styles["Default Character Font"]
-        hs.unhide_when_used = True
-        hs.font.color.rgb = docx.shared.RGBColor(0x05, 0x63, 0xC1)
-        hs.font.underline = True
-        del hs
-
-    return "Hyperlink"
-
-
-def add_hyperlink(
-    paragraph: Paragraph,
-    text: str,
-    url: str,
-) -> docx.oxml.shared.OxmlElement:
-    """Create a hyperlink object and add it to the paragraph."""
-
-    # This gets access to the document.xml.rels file and gets a new relation id value
-    part = paragraph.part
-    r_id = part.relate_to(
-        url,
-        docx.opc.constants.RELATIONSHIP_TYPE.HYPERLINK,
-        is_external=True,
-    )
-
-    # Create the w:hyperlink tag and add needed values
-    hyperlink = docx.oxml.shared.OxmlElement("w:hyperlink")
-    hyperlink.set(
-        docx.oxml.shared.qn("r:id"),
-        r_id,
-    )
-
-    # Create a new run object (a wrapper over a 'w:r' element)
-    new_run = docx.text.run.Run(docx.oxml.shared.OxmlElement("w:r"), paragraph)
-    new_run.text = text
-
-    # Set the run's style to the builtin hyperlink style, defining it if necessary
-    new_run.style = get_or_create_hyperlink_style(part.document)
-
-    # Join all the xml elements together
-    hyperlink.append(new_run._element)  # noqa: SLF001
-    paragraph._p.append(hyperlink)  # noqa: SLF001
-    return hyperlink
-
-
-Paragraph.add_hyperlink = add_hyperlink
 
 
 class BasicRenderPersonalSection(ResumeRenderPersonalBase):
@@ -101,7 +26,7 @@ class BasicRenderPersonalSection(ResumeRenderPersonalBase):
         log.debug("Initializing personal basic render object")
         super().__init__(document, personal, settings)
 
-    def _contact_info(self) -> None: #noqa: C901
+    def _contact_info(self) -> None:  # noqa: C901
         """Render the contact info section."""
 
         log.debug("Rendering contact info section")
@@ -131,7 +56,7 @@ class BasicRenderPersonalSection(ResumeRenderPersonalBase):
             _paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
 
         if _render_email:
-            _paragraph.add_hyperlink(_info.email, f"mailto: {_info.email}")
+            add_hyperlink(_paragraph, _info.email, f"mailto: {_info.email}")
             _has_content = True
 
         if _render_phone:
@@ -160,25 +85,25 @@ class BasicRenderPersonalSection(ResumeRenderPersonalBase):
         _has_content = False
 
         if _websites.github and self.settings.github:
-            _paragraph.add_hyperlink("GitHub", _websites.github)
+            add_hyperlink(_paragraph, "GitHub", _websites.github)
             _has_content = True
 
         if _websites.linkedin and self.settings.linkedin:
             if _has_content:
                 _paragraph.add_run(" | ")
-            _paragraph.add_hyperlink("LinkedIn", _websites.linkedin)
+            add_hyperlink(_paragraph, "LinkedIn", _websites.linkedin)
             _has_content = True
 
         if _websites.website and self.settings.website:
             if _has_content:
                 _paragraph.add_run(" | ")
-            _paragraph.add_hyperlink("Website", _websites.website)
+            add_hyperlink(_paragraph, "Website", _websites.website)
             _has_content = True
 
         if _websites.twitter and self.settings.twitter:
             if _has_content:
                 _paragraph.add_run(" | ")
-            _paragraph.add_hyperlink("X/Twitter", _websites.twitter)
+            add_hyperlink(_paragraph, "X/Twitter", _websites.twitter)
 
         if _has_content:
             _paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
